oss-fuzz: Fix undefined-behavior
Analyzer run time does not like A -= B or A +=B even if A and B are uint32_t as there is potential
for overflow.

Analyzer run time does not like overflow when doing coap_tick_t - coap_tick_t.

Analyzer run time does not like (uint8_t << 24) as uint8_t is promoted as int, not u_int.

Analyzer run time does not like while(A--) {} even if A is uint32_t, as A is always decremented on
the test, even if the value of A is 0.

Fix potential overflow when analyzing an asn.1 tag.

Fix potential overflow when doing decode_segment() on a % hex value.

+++ src/coap_block.c
@@ -3611,7 +3611,7 @@ derive_cbor_value(const uint8_t **bp, size_t rem_len) {
   }
   if (rem_len < 4)
     return (uint32_t)-1;
-  value = **bp << 24;
+  value = (uint32_t)(**bp) << 24;
   (*bp)++;
   value |= **bp << 16;
   (*bp)++;
